+++ src/prelude.rs
@@ -1,0 +1,5 @@
+pub use crate::log::{Logger, SilenceLevel};
+pub use crate::memory::AddressSpace;
+pub use crate::registers::Registers;
+pub use crate::test::TestConfig;
+pub use crate::{cpu, open_rom, open_symfile, run_tests};